Added a block which rises when the "held" variable is set to true, once there's a flashlight we can set it to true with the raycast.
Also, when trying to figure out how to get it to hold itself up for a few seconds in case we want it to press a button on the ceiling or something, I accidentally made a block that charges up for a jump until released, which is pretty neat, so I have that object too.

I haven't tested any scaling or anything, likely it'll work fine though.